requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-cors-headers==3.4.0
django-redis==4.12.1
djangorestframework==3.11.1
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
toml==0.10.1
